requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2  # Excel文件处理
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
